requirements.txt
slack_bolt
openai
solana-py
requests
streamlit
dotenv
